+++ release-notes.md
@@ -4,7 +4,7 @@ for [Chrome extension to for Bible.com](README.md) by [@Matei Nicolae](https://n
 
 💚 **Love this extension?** Share [feedback](https://chromewebstore.google.com/detail/project-verses-from-bible/fklnkmnlobkpoiifnbnemdpamheoanpj) and help us make it even better + helps others to find this resource.
 
-## 2.8.0 (🔜 coming soon)
+## 2.8.0 (2025-04-07)
 
 - [x] in projector page if there are checkboxes from live text "[x] / [ ]", allow to 'check/uncheck' (+ improve UX for them)
 - [x] when add new ref using short ref (ex. "2:4") - add it to the list using existing book short name (ex. "Exo 2:4"), search for book name in list